fix: remove config singleton pattern
