+++ packages/server-core/src/adapters/http/controllers/workflows/openid/index.ts
@@ -5,5 +5,4 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './handlers';
 export * from './module';
