+++ packages/server-core/src/adapters/http/swagger/type.ts
@@ -6,5 +6,5 @@
  */
 
 export type SwaggerOptions = {
-    baseURL?: string
+    baseURL: string
 };
