+++ packages/server-core/src/app/modules/config/index.ts
@@ -7,3 +7,4 @@
 
 export * from './constants';
 export * from './module';
+export * from './types';
