+++ packages/server-core/src/app/modules/config/read/module.ts
