--- packages/server-core/src/config/index.ts
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) 2022.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-export * from './constants';
-export * from './singleton';
-export * from './parse';
-export * from './types';
-export * from './normalize';
-export * from './read';
